Capstone 3, team 3 - change Longs to ints

+++ 09-Capstone/student/java/tenmo-client/src/main/java/com/techelevator/tenmo/models/Account.java
@@ -4,8 +4,8 @@ import java.math.BigDecimal;
 
 public class Account {
 	
-	private Long accountId;
-	private Long userId;
+	private int accountId;
+	private int userId;
 	private BigDecimal balance;
 	
 	
@@ -14,7 +14,7 @@ public class Account {
 	public Account() {
 		
 	}
-	public Account(Long accountId, Long userId, BigDecimal balance) {
+	public Account(int accountId, int userId, BigDecimal balance) {
 		this.accountId = accountId;
 		this.userId = userId;
 		this.balance = balance;
@@ -22,10 +22,10 @@ public class Account {
 	
 	
 	
-	public Long getAccountId() {
+	public int getAccountId() {
 		return accountId;
 	}
-	public Long getUserId() {
+	public int getUserId() {
 		return userId;
 	}
 	public BigDecimal getBalance() {
